chore(tasks): reorganize task names
