+++ tasks/tasks/dev/permissioned.ts
@@ -8,12 +8,12 @@ import {
   logConstructorArgs,
   logDeploymentConfig,
   readAndValidateParams,
-} from '../../../common/deploy/helpers';
-import { getNetworkKey } from '../../../common/utils';
-import { sedaScope } from '../../../index';
+} from '../../common/deploy/helpers';
+import { getNetworkKey } from '../../common/utils';
+import { sedaScope } from '../../index';
 
 sedaScope
-  .task('deploy:dev:permissioned', 'Deploys the Permissioned SEDA contract (only for testing)')
+  .task('dev:deploy:permissioned', 'Deploys the Permissioned SEDA contract (only for testing)')
   .addOptionalParam('maxReplicationFactor', 'The maximum replication factor', undefined, types.int)
   .addOptionalParam('params', 'The parameters file to use', undefined, types.string)
   .addFlag('reset', 'Replace existing deployment files')
