+++ tasks/tasks/dev/resettableProver.ts
@@ -1,11 +1,11 @@
 import { types } from 'hardhat/config';
 import type { HardhatRuntimeEnvironment } from 'hardhat/types';
 
-import { sedaScope } from '../../../index';
-import { deployProver } from '../proverBase';
+import { sedaScope } from '../../index';
+import { deployProver } from '../deploy/proverBase';
 
 sedaScope
-  .task('deploy:dev:prover-reset', 'Deploys the Secp256k1ProverResettable contract (only for testing)')
+  .task('dev:deploy:prover', 'Deploys the Secp256k1ProverResettable contract (only for testing)')
   .addParam('params', 'The parameters file to use', undefined, types.string)
   .addOptionalParam('maxBatchAge', 'The maximum allowed age difference between batches', undefined, types.bigint)
   .addOptionalParam('feeManagerAddress', 'The address of the FeeManager contract', undefined, types.string)
